requirements.txt
boto3>=1.34.0
plaid-python>=15.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
python-dotenv>=1.0.0
requests>=2.31.0
pynacl>=1.5.0